requirements.txt
flask==3.0.0
flask-cors==4.0.0
firebase-admin>=6.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-mock==3.12.0

